Split sample.py module
